Added Dye Util class to contain all utilities needed for working with MC dyes Fixed Light Gray Dye (Silver) using bad oreDict name Fixed all usages of dyes in recipes to use correct oreDict name
